Melhorias

+++ src/main/java/gov/caixa/invest/dto/SimulacaoListItem.java
@@ -1,6 +1,7 @@
 package gov.caixa.invest.dto;
 
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 
@@ -8,7 +9,7 @@ public class SimulacaoListItem {
     public Long id;
     public Long clienteId;
     public String produto;
-    public double valorInvestido;
+    public BigDecimal valorInvestido;
     public double valorFinal;
     public int prazoMeses;
     public OffsetDateTime dataSimulacao;
